feat(footer): BUGGED add CC-BY-4 default to footer.
Bugged in that while included in DOM, doesn't visually display.
Probably a styling issue, have not yet figured it out.

This language is verbatim that from Creative Commons site, except without
the otherwise noted hyperlink, since its not clear there's somewhere for that to link to yet.

Differs from status quo apereo.org in that license version++ and shorter form eliding who is doing the licensing. I figure if anyone's getting this right CC is so we just don't need the longer form. Also I love the idea of direct licensing from authors to consumers and not via Apereo re-licensing. inbound==outbound.